move to sub package the creation of the model
