+++ measurements_model/model_execution/extract_model.py
@@ -2,12 +2,12 @@ import joblib
 
 from measurements_model.config import TRAIN_SET_PATH, FULL_PREPROCESSED_DATASET_PATH, IDLE_DIR_PATH, TEST_SET_PATH, \
     ProcessColumns
-from measurements_model.dataset_pipeline_executor import DatasetPipelineExecutor
+from measurements_model.model_execution.dataset_pipeline_executor import DatasetPipelineExecutor
 from measurements_model.dataset_processing.feature_selection.process_and_system_no_hardware_feature_selector import \
     ProcessAndSystemNoHardware
 from measurements_model.dataset_processing.split_data.regular_spliter import RegularDatasetSplitter
-from measurements_model.main_model import BestModelConfig
-from measurements_model.main_model_configuration import ALL_MEASUREMENTS_DIRS_PATH, MODEL_FILE_NAME
+from measurements_model.model_execution.main_model import BestModelConfig
+from measurements_model.model_execution.main_model_configuration import ALL_MEASUREMENTS_DIRS_PATH, MODEL_FILE_NAME
 
 if __name__ == '__main__':
     feature_selector = ProcessAndSystemNoHardware()
